planner/requirements.txt
paho-mqtt
influxdb-client
numpy
